refactor: rename db-manager to background, move all other background tasks there
This leaves the server with the sole responsibility of serving GraphQL requests,
with in-memory caching an atomic part of this.
