+++ package.json
@@ -21,35 +21,35 @@
     "mainnet:up": "yarn mainnet:stack && docker compose -p mainnet logs -f",
     "mainnet:down": "docker compose -p mainnet down --remove-orphans",
     "mainnet:server": "CONTEXT=$PWD NETWORK=mainnet . ./scripts/export_env.sh && yarn workspace @cardano-graphql/server start",
-    "mainnet:db-manager": "CONTEXT=$PWD NETWORK=mainnet . ./scripts/export_env.sh && yarn workspace @cardano-graphql/api-cardano-db-hasura db-manager",
+    "mainnet:background": "CONTEXT=$PWD NETWORK=mainnet . ./scripts/export_env.sh && yarn workspace @cardano-graphql/api-cardano-db-hasura background",
     "preprod:stack": "CONTEXT=$PWD NETWORK=preprod . ./scripts/export_env.sh && docker compose -p preprod up --build -d",
     "preprod:dev": "CONTEXT=$PWD NETWORK=preprod . ./scripts/export_env.sh && yarn service-dependencies -p preprod up --build",
     "preprod:up": "yarn preprod:stack && docker compose -p preprod logs -f",
     "preprod:down": "docker compose -p preprod down --remove-orphans",
     "preprod:server": "CONTEXT=$PWD NETWORK=preprod . ./scripts/export_env.sh && yarn workspace @cardano-graphql/server start",
-    "preprod:db-manager": "CONTEXT=$PWD NETWORK=preprod . ./scripts/export_env.sh && yarn workspace @cardano-graphql/api-cardano-db-hasura db-manager",
+    "preprod:background": "CONTEXT=$PWD NETWORK=preprod . ./scripts/export_env.sh && yarn workspace @cardano-graphql/api-cardano-db-hasura background",
     "preview:stack": "CONTEXT=$PWD NETWORK=preview . ./scripts/export_env.sh && docker compose -p preview up --build -d",
     "preview:dev": "CONTEXT=$PWD NETWORK=preview . ./scripts/export_env.sh && yarn service-dependencies -p preview up --build",
     "preview:up": "yarn preview:stack && docker compose -p preview logs -f",
     "preview:down": "docker compose -p preview down --remove-orphans",
     "preview:server": "CONTEXT=$PWD NETWORK=preview . ./scripts/export_env.sh && yarn workspace @cardano-graphql/server start",
-    "preview:db-manager": "CONTEXT=$PWD NETWORK=preview . ./scripts/export_env.sh && yarn workspace @cardano-graphql/api-cardano-db-hasura db-manager",
+    "preview:background": "CONTEXT=$PWD NETWORK=preview . ./scripts/export_env.sh && yarn workspace @cardano-graphql/api-cardano-db-hasura background",
     "testnet:stack": "CONTEXT=$PWD NETWORK=testnet . ./scripts/export_env.sh && docker compose -p testnet up --build -d",
     "testnet:dev": "CONTEXT=$PWD NETWORK=testnet . ./scripts/export_env.sh && yarn service-dependencies -p testnet up --build",
     "testnet:up": "yarn testnet:stack && docker compose -p testnet logs -f",
     "testnet:down": "docker compose -p testnet down --remove-orphans",
     "testnet:server": "CONTEXT=$PWD NETWORK=testnet . ./scripts/export_env.sh && yarn workspace @cardano-graphql/server start",
-    "testnet:db-manager": "CONTEXT=$PWD NETWORK=testnet . ./scripts/export_env.sh && yarn workspace @cardano-graphql/api-cardano-db-hasura db-manager",
+    "testnet:background": "CONTEXT=$PWD NETWORK=testnet . ./scripts/export_env.sh && yarn workspace @cardano-graphql/api-cardano-db-hasura background",
     "vasil-dev:stack": "CONTEXT=$PWD NETWORK=vasil-dev . ./scripts/export_env.sh && docker compose -p vasil-dev up --build -d",
     "vasil-dev:dev": "CONTEXT=$PWD NETWORK=vasil-dev . ./scripts/export_env.sh && yarn service-dependencies -p vasil-dev up --build",
     "vasil-dev:up": "yarn vasil-dev:stack && docker compose -p vasil-dev logs -f",
     "vasil-dev:down": "docker compose -p vasil-dev down --remove-orphans",
     "vasil-dev:server": "CONTEXT=$PWD NETWORK=vasil-dev . ./scripts/export_env.sh && yarn workspace @cardano-graphql/server start",
-    "vasil-dev:db-manager": "CONTEXT=$PWD NETWORK=vasil-dev . ./scripts/export_env.sh && yarn workspace @cardano-graphql/api-cardano-db-hasura db-manager",
+    "vasil-dev:background": "CONTEXT=$PWD NETWORK=vasil-dev . ./scripts/export_env.sh && yarn workspace @cardano-graphql/api-cardano-db-hasura background",
     "lint": "yarn workspaces run lint",
     "loadtest:byron-staging": "artillery run test/loadtest/byron-staging-config.yml",
     "publish-packages": "yarn workspaces run publish",
-    "service-dependencies": "docker-compose config --services | grep -vE \"server|db-manager\" | xargs docker-compose",
+    "service-dependencies": "docker-compose config --services | grep -vE \"server|background\" | xargs docker-compose",
     "test": "yarn workspaces run test"
   },
   "contributors": [
