--- packages/api-cardano-db-hasura/src/db-manager/index.ts
@@ -1,2 +1,0 @@
-export * from './Db'
-export * from './HasuraManagementClient'
