+++ packages/api-cardano-db-hasura/src/index.ts
@@ -2,10 +2,11 @@ export * from './AssetMetadata'
 export * from './CardanoNodeClient'
 export * from './ChainFollower'
 export * from './Config'
+export * from './Db'
 export * from './executableSchema'
 export * from './graphql_types'
 export * from './HasuraClient'
-export * from './db-manager'
+export * from './HasuraBackgroundClient'
 export * from './MetadataClient'
 export * from './typeAliases'
 export * from './Worker'
